Watz ein Circuits
Added 1.7 Watz Reactor
Removed old Watz stuff
Added circuits
Added welder and solderer
Added more Brick variants

Updated nuke and oil models

Fixed nuke positioning
Fixed fluid transparency when inside fluid blocks
Fixed Crafting handles not working in 2x2

+++ src/main/java/com/hbm/blocks/bomb/DetCord.java
@@ -78,7 +78,7 @@ public class DetCord extends Block implements IBomb {
 				bf.posX = pos.getX() + 0.5;
 				bf.posY = pos.getY() + 0.5;
 				bf.posZ = pos.getZ() + 0.5;
-				bf.destructionRange = (int) 130;
+				bf.destructionRange = 130;
 				world.spawnEntity(bf);
 				if(BombConfig.enableNukeClouds) {
 					EntityNukeTorex.statFacBale(world, pos.getX() + 0.5, pos.getY() + 5, pos.getZ() + 0.5, 130F);
